Rename Pointer Handlers to Input Handlers (as a concept)
This is a documentation change to alleviate some confusion that we've
seen during the Tech Preview period.

It doesn't make sense to actually rename the base class though, because
it is intended to handle QQuickPointerEvents, not QEvents.  The reason
for that is that refactoring the QEvent hierarchy has to wait until
Qt 6.  So maybe in Qt 6 we can remove QQuickPointerEvent and have
a QQuickInputHandler base class which handles QInputEvents; but for
now, this conceptual renaming seems about as far as we can go.

Task-number: QTBUG-66651
Change-Id: I84a41dc282c480d08f4d4a0d9a857e37e074aa7a
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@qt.io>