tst_qquickpositioners::createView: replace qDebugs, fix warning
We use qCDebug(lcTests) to avoid the output clutter unless the test
fails, then the CI system will re-run it with all logging categories
enabled.  The "1-2-3-4" count was anyway not very informative.

As usual there was a -Wunused-result warning about ignoring the
return from qWaitForExposed, so now we don't ignore it.  (Not sure
if it ought to be fatal though.)

Change-Id: I48f43083cbbcf27c81f0ac951b03d7b8b76e5300
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>