Add run time assertion for local handling
Commit f44782d0cdbdb800d9c31d5aff712fbf29d52edc fixed the missing call
context creation that would lead to
language/statements/break/S12.8_A2.js failing. It wouldn't always fail
as the invalid cast from the global context to a call context would
access memory that happens to be available just for the test. An
assertion however will not require us to rely on memory setup.

Change-Id: I131a2242004cd5e4d518e58cc9f6a79037f962d2
Reviewed-by: Lars Knoll <lars.knoll@qt.io>