QQmlDelegateModelItem: set row and column explicit
Whenever we set a model index on a model item, set row and
column explicit as well. We used to calculate row and column
by calling out to QQmlDelegateModel from within
QQmlDelegateModelItem. But we want to remove the dependency
to QQmlDelegateModel as much as possible, to be able to reuse
QQmlDelegateModelItem also from other model classes (first and
foremost the upcoming QQmlTableInstanceModel).

Change-Id: Ib80b31250010b2611ba1715d3324145117add865
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>