QQmlInstanceModel: add virtual function to get abstractItemModel
Add a new virtual function to query the model if it's backed by
a QAbstractItemModel. This will be needed later when we create
another model (QQmlTableInstanceModel) that also inherits
QQmlInstanceModel.

By having this virtual function, we don't have to check and special
case what kind of model is assigned to QQuickTableView, as long
as it has a QAIM.

Change-Id: I4080fd601ca1988c5fe448d68bdd156740403c60
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>