Quick tests: remember positions of stationary points
We need Items to receive accurate positions for stationary touch
points: that is, the last-known position from the previous touch
event.  (And we hope that all actual touch-capable platforms also send
proper QPA events with correct positions for stationary points.
We assert that it's a bug if they don't.)

As explained in qtbase 7cef4b6463fdb73ff602ade64b222333dd23e46c, it's
OK to retain a copy of a QTest::QTouchEventSequence for this purpose,
so that the QMap<int, QTouchEvent::TouchPoint> previousPoints will not
be discarded between events.

Change-Id: I12a04323b0d82dc325348c0fa3e067a4b63f57e9
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>