QQWindowPriv::sendFilteredPointerEvent: don't clear hasFiltered
It should be the caller's responsibility to do that when needed.
In cases when sendFilteredPointerEvent is being called repeatedly
in a loop, it doesn't make sense to allow the same ancestor item
to filter the same event more than once.  For example Flickable will
decide to steal the grab the first time it sees a mouse movement
that exceeds the drag threshold, but it is supposed to wait until
the next event to actually steal the grab.  But if it gets a chance
to filter twice, it will steal the grab within the context of the
delivery of one event.

5cb76fb3704947cfc4d575695b137460ecc8bbd9 took care of repeated
filtering in case of two handlers belonging to the same Item; from one
side, this patch is more general than that.  But another purpose of
that patch was also to prevent delivery to any more Pointer Handlers
if an Item parent has already filtered the event.  hasFiltered
remembers which parents had a chance to filter;
sendFilteredPointerEventResult remembers what each of them returned
from the filter function.  It would be nice if we could avoid needing
both of these data structures.

Task-number: QTBUG-64846
Change-Id: I4d26899224e8531ff3d74dbb35994a26e5e94763
Reviewed-by: Frederik Gladhorn <frederik.gladhorn@qt.io>