Call clearGrabbers each time we get a press event
This is important in order for passive grabbers to be in the same order
as if the points were pressed at the same time.

In our case, the problem occurred when we had a single-point DragHandler
together with a two-finger PinchHandler:

* One finger was pressed and moved
  => DragHandler called setPassiveGrab()
     => point0->passiveGrabbers: [DragHandler]
* A second finger was pressed and moved
  => PinchHandler called setPassiveGrab() for both points
     => point0->passiveGrabbers: [DragHandler,PinchHandler]
     => point1->passiveGrabbers: [PinchHandler]

So then as one keeps on dragging the *two* fingers, the DragHandler will
get the chance to do an exclusive grab first, (since its the first listed
passive grabber of point0), and the PinchHandler won't get the opportunity
to grab. This is not expected since their declaration order implies that
the PinchHandler should get a chance to grab first.

Change-Id: I4e82ed186eeb5bf1dae1679d393e5563072175d1
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>