TapHandler: enforce drag threshold for timeHeld and longPressed
As the docs say, timeHeld is the amount of time that a pressed point
has been held, without moving beyond the drag threshold.  Likewise
the longPressed signal should not be omitted if the point has moved
beyond the drag threshold.  In typical interactions these are relevant
for long-pressing in one place, not while moving around.

Task-number: QTBUG-65012
Change-Id: I3236def5de4d70f8d174362972a76e680452f018
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>