Flickable: never react to a QMouseEvent with no buttons pressed
If a Flickable contains an Item with a HoverHandler and a DragHandler,
and you drag it and release it, then during the next mouse move events,
since Flickable is able to filter those while they are being sent to
the HoverHandler, it would grab the mouse.  Suddenly you'd be flicking
even with no mouse buttons pressed.  I think I've seen similar behavior
in other scenarios too, at some point.

Change-Id: If79c8af2b62dc69f085513e0b7c8bf9b4d504572
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>