Reduce memory usage caused by strings with AOT cache files
For AOT loaded cache files, the CompiledData::Unit we produce is
ephemeral by design. That same applies to the strings, where we don't
really need to write out a new string table but we might as well just
keep the QStringList around. That avoids copying all the string data and
saves another 21K RAM with the gallery. It might even save a little more
if some of the new strings are truly shared via implicit QString
sharing. Also their "extraction" via stringAt(index) becomes as fast as
for the strings that come from mmap'ed memory - with no copying
involved.

Task-number: QTBUG-69588
Change-Id: I06cf0f083e3b8d3c5dbabb22beb0711f88fc8692
Reviewed-by: Lars Knoll <lars.knoll@qt.io>