Remove parallax example
The example was part of the views example, except it was not even listed
in the application, so it would not be discovered by anyone in the first
place. Together with 3rd party assets that would need to be documented,
and it being broken after samegame was moved, there is no reason to keep
this.

Task-number: QTBUG-69660
Change-Id: Ib701bd8987fd1d8473e1073bf4d701dbb087fdba
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
Reviewed-by: Kai Koehne <kai.koehne@qt.io>